requirements.txt
mc.py
aiogram==2.25.1
beautifulsoup4
requests
lxml
loguru
selenium
python-mpv
py-cpuinfo
psutil
langdetect
gtts
python-magic